chore(dev-dashboard): move component to common - remove redundant file
