+++ ui/apps/dev-dashboard/src/components/common/index.ts
@@ -1,1 +1,2 @@
 export * from './card-wrapper';
+export * from './dev-message-form';
